+++ main/index.ts
@@ -46,4 +46,8 @@ export {
 } from '../ipc/schema/packageJson';
 
 export { type OverwriteOptions } from '../ipc/serialPort';
-export { type OpenAppOptions } from '../ipc/openWindow';
+export type { OpenAppOptions } from '../ipc/openWindow';
+export {
+    isOpenAppOptionsDevicePort,
+    isOpenAppOptionsDeviceSN,
+} from '../ipc/openWindow';